requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
easydict==1.13
flatbuffers==25.2.10
fonttools==4.59.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
moviepy==2.2.1
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy==1.26.4
onnx==1.18.0
onnxruntime-gpu==1.18.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
prettytable==3.16.0
proglog==0.1.12
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
